refactor: Split token api into smaller pieces
Refactor the token API splitting it into smaller pieces that allow
easier maintainability.
