+++ src/api/v3/auth/token/common.rs
@@ -13,165 +13,21 @@
 // SPDX-License-Identifier: Apache-2.0
 
 use crate::api::common;
-use crate::api::error::{KeystoneApiError, TokenError};
 use crate::api::types::ProjectBuilder;
-use crate::api::v3::auth::token::types::{Token, TokenBuilder, UserBuilder};
-use crate::api::v3::role::types::Role;
+use crate::api::v3::auth::token::types::AuthRequest;
+use crate::api::{
+    Scope,
+    common::{find_project_from_scope, get_domain},
+    error::KeystoneApiError,
+};
+use crate::auth::{AuthenticatedInfo, AuthzInfo};
 use crate::identity::IdentityApi;
 use crate::keystone::ServiceState;
-use crate::resource::{
-    ResourceApi,
-    types::{Domain, Project},
-};
-use crate::token::Token as ProviderToken;
+use crate::resource::types::{Domain, Project};
+use crate::token::TokenApi;
 
-impl Token {
-    pub async fn from_provider_token(
-        state: &ServiceState,
-        token: &ProviderToken,
-    ) -> Result<Token, KeystoneApiError> {
-        let mut response = TokenBuilder::default();
-        let mut project: Option<Project> = token.project().cloned();
-        let mut domain: Option<Domain> = token.domain().cloned();
-        response.audit_ids(token.audit_ids().clone());
-        response.methods(token.methods().clone());
-        response.expires_at(*token.expires_at());
-
-        let user = if let Some(user) = token.user() {
-            user
-        } else {
-            &state
-                .provider
-                .get_identity_provider()
-                .get_user(state, token.user_id())
-                .await
-                .map_err(KeystoneApiError::identity)?
-                .ok_or_else(|| KeystoneApiError::NotFound {
-                    resource: "user".into(),
-                    identifier: token.user_id().clone(),
-                })?
-        };
-
-        let user_domain = common::get_domain(state, Some(&user.domain_id), None::<&str>).await?;
-
-        let mut user_response: UserBuilder = UserBuilder::default();
-        user_response.id(user.id.clone());
-        user_response.name(user.name.clone());
-        if let Some(val) = user.password_expires_at {
-            user_response.password_expires_at(val);
-        }
-        user_response.domain(user_domain.clone());
-        response.user(user_response.build().map_err(TokenError::from)?);
-
-        if let Some(roles) = token.roles() {
-            response.roles(
-                roles
-                    .clone()
-                    .into_iter()
-                    .map(Into::into)
-                    .collect::<Vec<Role>>(),
-            );
-        }
-
-        match token {
-            ProviderToken::Unscoped(_token) => {}
-            ProviderToken::DomainScope(token) => {
-                if domain.is_none() {
-                    domain = Some(
-                        common::get_domain(state, Some(&token.domain_id), None::<&str>).await?,
-                    );
-                }
-            }
-            ProviderToken::ProjectScope(token) => {
-                if project.is_none() {
-                    project = Some(
-                        state
-                            .provider
-                            .get_resource_provider()
-                            .get_project(state, &token.project_id)
-                            .await
-                            .map_err(KeystoneApiError::resource)?
-                            .ok_or_else(|| KeystoneApiError::NotFound {
-                                resource: "project".into(),
-                                identifier: token.project_id.clone(),
-                            })?,
-                    );
-                }
-            }
-            ProviderToken::ApplicationCredential(token) => {
-                if project.is_none() {
-                    project = Some(
-                        state
-                            .provider
-                            .get_resource_provider()
-                            .get_project(state, &token.project_id)
-                            .await
-                            .map_err(KeystoneApiError::resource)?
-                            .ok_or_else(|| KeystoneApiError::NotFound {
-                                resource: "project".into(),
-                                identifier: token.project_id.clone(),
-                            })?,
-                    );
-                }
-            }
-            ProviderToken::FederationUnscoped(_token) => {}
-            ProviderToken::FederationDomainScope(token) => {
-                if domain.is_none() {
-                    domain = Some(
-                        common::get_domain(state, Some(&token.domain_id), None::<&str>).await?,
-                    );
-                }
-            }
-            ProviderToken::FederationProjectScope(token) => {
-                if project.is_none() {
-                    project = Some(
-                        state
-                            .provider
-                            .get_resource_provider()
-                            .get_project(state, &token.project_id)
-                            .await
-                            .map_err(KeystoneApiError::resource)?
-                            .ok_or_else(|| KeystoneApiError::NotFound {
-                                resource: "project".into(),
-                                identifier: token.project_id.clone(),
-                            })?,
-                    );
-                }
-            }
-            ProviderToken::Restricted(token) => {
-                if project.is_none() {
-                    project = Some(
-                        state
-                            .provider
-                            .get_resource_provider()
-                            .get_project(state, &token.project_id)
-                            .await
-                            .map_err(KeystoneApiError::resource)?
-                            .ok_or_else(|| KeystoneApiError::NotFound {
-                                resource: "project".into(),
-                                identifier: token.project_id.clone(),
-                            })?,
-                    );
-                }
-            }
-        }
-
-        if let Some(domain) = domain {
-            response.domain(domain.clone());
-        }
-        if let Some(project) = project {
-            response.project(
-                get_project_info_builder(state, &project, &user_domain)
-                    .await?
-                    .build()
-                    .map_err(TokenError::from)?,
-            );
-        }
-        Ok(response.build().map_err(TokenError::from)?)
-    }
-}
-
-async fn get_project_info_builder(
+/// Get the ProjectBuilder for the given Project.
+pub(super) async fn get_project_info_builder(
     state: &ServiceState,
     project: &Project,
     user_domain: &Domain,
@@ -189,64 +45,155 @@ async fn get_project_info_builder(
     Ok(project_response)
 }
 
+/// Authenticate the user ignoring any scope information. It is important not to expose any
+/// hints that user, project, domain, etc might exist before we have authenticated them by
+/// taking different amount of time in case of certain validations.
+pub(super) async fn authenticate_request(
+    state: &ServiceState,
+    req: &AuthRequest,
+) -> Result<AuthenticatedInfo, KeystoneApiError> {
+    let mut authenticated_info: Option<AuthenticatedInfo> = None;
+    for method in req.auth.identity.methods.iter() {
+        if method == "password" {
+            if let Some(password_auth) = &req.auth.identity.password {
+                let req = password_auth.user.clone().try_into()?;
+                authenticated_info = Some(
+                    state
+                        .provider
+                        .get_identity_provider()
+                        .authenticate_by_password(state, req)
+                        .await?,
+                );
+            }
+        } else if method == "token"
+            && let Some(token) = &req.auth.identity.token
+        {
+            let mut authz = state
+                .provider
+                .get_token_provider()
+                .authenticate_by_token(state, &token.id, Some(false), None)
+                .await?;
+            // Resolve the user
+            authz.user = Some(
+                state
+                    .provider
+                    .get_identity_provider()
+                    .get_user(state, &authz.user_id)
+                    .await
+                    .map(|x| {
+                        x.ok_or_else(|| KeystoneApiError::NotFound {
+                            resource: "user".into(),
+                            identifier: authz.user_id.clone(),
+                        })
+                    })??,
+            );
+            authenticated_info = Some(authz);
+
+            {}
+        }
+    }
+    authenticated_info
+        .ok_or(KeystoneApiError::Unauthorized)
+        .and_then(|authn| {
+            authn.validate()?;
+            Ok(authn)
+        })
+}
+
+/// Build the AuthZ information from the request
+///
+/// # Arguments
+///
+/// * `state` - The service state
+/// * `req` - The Request
+///
+/// # Result
+///
+/// * `Ok(AuthzInfo)` - The AuthZ information
+/// * `Err(KeystoneApiError)` - The error
+pub(super) async fn get_authz_info(
+    state: &ServiceState,
+    req: &AuthRequest,
+) -> Result<AuthzInfo, KeystoneApiError> {
+    let authz_info = match &req.auth.scope {
+        Some(Scope::Project(scope)) => {
+            if let Some(project) = find_project_from_scope(state, scope).await? {
+                AuthzInfo::Project(project)
+            } else {
+                return Err(KeystoneApiError::Unauthorized);
+            }
+        }
+        Some(Scope::Domain(scope)) => {
+            if let Ok(domain) = get_domain(state, scope.id.as_ref(), scope.name.as_ref()).await {
+                AuthzInfo::Domain(domain)
+            } else {
+                return Err(KeystoneApiError::Unauthorized);
+            }
+        }
+        Some(Scope::System(_scope)) => {
+            todo!()
+        }
+        None => AuthzInfo::Unscoped,
+    };
+    authz_info.validate()?;
+    Ok(authz_info)
+}
+
 #[cfg(test)]
 mod tests {
     use sea_orm::DatabaseConnection;
     use std::sync::Arc;
 
-    use crate::api::v3::auth::token::types::Token;
-    use crate::api::v3::role::types::Role;
-    use crate::assignment::{
-        MockAssignmentProvider,
-        types::{Assignment, AssignmentType, Role as ProviderRole, RoleAssignmentListParameters},
-    };
+    use crate::api::KeystoneApiError;
 
+    use crate::auth::AuthenticatedInfo;
     use crate::config::Config;
-    use crate::identity::{MockIdentityProvider, types::UserResponse};
+    use crate::identity::{
+        MockIdentityProvider,
+        types::{UserPasswordAuthRequest, UserResponse},
+    };
     use crate::keystone::Service;
     use crate::policy::MockPolicyFactory;
     use crate::provider::Provider;
-    use crate::resource::{
-        MockResourceProvider,
-        types::{Domain, Project},
-    };
-    use crate::token::{
-        DomainScopePayload, ProjectScopePayload, Token as ProviderToken, UnscopedPayload,
-    };
+
+    use crate::token::MockTokenProvider;
+
+    use super::super::types::*;
+    use super::*;
 
     #[tokio::test]
-    async fn test_from_unscoped() {
+    async fn test_authenticate_request_password() {
+        let config = Config::default();
+        let auth_info = AuthenticatedInfo::builder()
+            .user_id("uid")
+            .user(UserResponse {
+                id: "uid".to_string(),
+                domain_id: "udid".into(),
+                enabled: true,
+                ..Default::default()
+            })
+            .build()
+            .unwrap();
+        let auth_clone = auth_info.clone();
         let mut identity_mock = MockIdentityProvider::default();
         identity_mock
-            .expect_get_user()
-            .withf(|_, id: &'_ str| id == "bar")
-            .returning(|_, _| {
-                Ok(Some(UserResponse {
-                    id: "bar".into(),
-                    domain_id: "user_domain_id".into(),
-                    ..Default::default()
-                }))
-            });
+            .expect_authenticate_by_password()
+            .withf(|_, req: &UserPasswordAuthRequest| {
+                req.id == Some("uid".to_string())
+                    && req.password == "pwd"
+                    && req.name == Some("uname".to_string())
+            })
+            .returning(move |_, _| Ok(auth_clone.clone()));
 
-        let mut resource_mock = MockResourceProvider::default();
-        resource_mock
-            .expect_get_domain()
-            .withf(|_, id: &'_ str| id == "user_domain_id")
-            .returning(|_, _| {
-                Ok(Some(Domain {
-                    id: "user_domain_id".into(),
-                    ..Default::default()
-                }))
-            });
         let provider = Provider::mocked_builder()
+            .config(config.clone())
             .identity(identity_mock)
-            .resource(resource_mock)
             .build()
             .unwrap();
 
         let state = Arc::new(
             Service::new(
-                Config::default(),
+                config,
                 DatabaseConnection::Disconnected,
                 provider,
                 MockPolicyFactory::new(),
@@ -254,167 +201,145 @@ mod tests {
             .unwrap(),
         );
 
-        let api_token = Token::from_provider_token(
-            &state,
-            &ProviderToken::Unscoped(UnscopedPayload {
-                user_id: "bar".into(),
-                ..Default::default()
-            }),
-        )
-        .await
-        .unwrap();
-        assert_eq!("bar", api_token.user.id);
-        assert_eq!(Some("user_domain_id"), api_token.user.domain.id.as_deref());
-        assert!(api_token.project.is_none());
-        assert!(api_token.domain.is_none());
-    }
-
-    #[tokio::test]
-    async fn test_from_domain_scoped() {
-        let mut identity_mock = MockIdentityProvider::default();
-        identity_mock
-            .expect_get_user()
-            .withf(|_, id: &'_ str| id == "bar")
-            .returning(|_, _| {
-                Ok(Some(UserResponse {
-                    id: "bar".into(),
-                    domain_id: "user_domain_id".into(),
-                    ..Default::default()
-                }))
-            });
-
-        let mut resource_mock = MockResourceProvider::default();
-        resource_mock
-            .expect_get_domain()
-            .returning(|_, id: &'_ str| {
-                Ok(Some(Domain {
-                    id: id.to_string(),
-                    ..Default::default()
-                }))
-            });
-        let provider = Provider::mocked_builder()
-            .identity(identity_mock)
-            .resource(resource_mock)
-            .build()
-            .unwrap();
-
-        let state = Arc::new(
-            Service::new(
-                Config::default(),
-                DatabaseConnection::Disconnected,
-                provider,
-                MockPolicyFactory::new(),
-            )
-            .unwrap(),
-        );
-
-        let api_token = Token::from_provider_token(
-            &state,
-            &ProviderToken::DomainScope(DomainScopePayload {
-                user_id: "bar".into(),
-                domain_id: "domain_id".into(),
-                ..Default::default()
-            }),
-        )
-        .await
-        .unwrap();
-
-        assert_eq!("bar", api_token.user.id);
-        assert_eq!(Some("user_domain_id"), api_token.user.domain.id.as_deref());
         assert_eq!(
-            Some("domain_id"),
-            api_token.domain.expect("domain scope").id.as_deref()
+            auth_info,
+            authenticate_request(
+                &state,
+                &AuthRequest {
+                    auth: AuthRequestInner {
+                        identity: Identity {
+                            methods: vec!["password".to_string()],
+                            password: Some(PasswordAuth {
+                                user: UserPassword {
+                                    id: Some("uid".to_string()),
+                                    password: "pwd".to_string(),
+                                    name: Some("uname".to_string()),
+                                    ..Default::default()
+                                },
+                            }),
+                            token: None,
+                        },
+                        scope: None,
+                    },
+                }
+            )
+            .await
+            .unwrap()
         );
-        assert!(api_token.project.is_none());
     }
 
     #[tokio::test]
-    async fn test_from_project_scoped() {
+    async fn test_authenticate_request_token() {
+        let config = Config::default();
+
+        let mut token_mock = MockTokenProvider::default();
+        token_mock
+            .expect_authenticate_by_token()
+            .withf(
+                |_, id: &'_ str, allow_expired: &Option<bool>, window: &Option<i64>| {
+                    id == "fake_token" && *allow_expired == Some(false) && window.is_none()
+                },
+            )
+            .returning(|_, _, _, _| {
+                Ok(AuthenticatedInfo::builder().user_id("uid").build().unwrap())
+            });
         let mut identity_mock = MockIdentityProvider::default();
         identity_mock
             .expect_get_user()
-            .withf(|_, id: &'_ str| id == "bar")
-            .returning(|_, _| {
+            .withf(|_, id: &'_ str| id == "uid")
+            .returning(|_, id: &'_ str| {
                 Ok(Some(UserResponse {
-                    id: "bar".into(),
+                    id: id.to_string(),
                     domain_id: "user_domain_id".into(),
+                    enabled: true,
                     ..Default::default()
                 }))
             });
 
-        let mut resource_mock = MockResourceProvider::default();
-        resource_mock
-            .expect_get_domain()
-            .returning(|_, id: &'_ str| {
-                Ok(Some(Domain {
-                    id: id.to_string(),
+        let provider = Provider::mocked_builder()
+            .config(config.clone())
+            .identity(identity_mock)
+            .token(token_mock)
+            .build()
+            .unwrap();
+
+        let state = Arc::new(
+            Service::new(
+                config,
+                DatabaseConnection::Disconnected,
+                provider,
+                MockPolicyFactory::new(),
+            )
+            .unwrap(),
+        );
+
+        assert_eq!(
+            AuthenticatedInfo::builder()
+                .user_id("uid")
+                .user(UserResponse {
+                    id: "uid".to_string(),
+                    domain_id: "user_domain_id".into(),
+                    enabled: true,
                     ..Default::default()
-                }))
-            });
-        resource_mock
-            .expect_get_project()
-            .returning(|_, id: &'_ str| {
-                Ok(Some(Project {
-                    id: id.to_string(),
-                    domain_id: "project_domain_id".into(),
-                    ..Default::default()
-                }))
-            });
-        let mut assignment_mock = MockAssignmentProvider::default();
-        assignment_mock.expect_list_role_assignments().returning(
-            |_, q: &RoleAssignmentListParameters| {
-                Ok(vec![Assignment {
-                    role_id: "rid".into(),
-                    role_name: Some("role_name".into()),
-                    actor_id: q.user_id.clone().unwrap(),
-                    target_id: q.project_id.clone().unwrap(),
-                    r#type: AssignmentType::UserProject,
-                    inherited: false,
-                }])
+                })
+                .build()
+                .unwrap(),
+            authenticate_request(
+                &state,
+                &AuthRequest {
+                    auth: AuthRequestInner {
+                        identity: Identity {
+                            methods: vec!["token".to_string()],
+                            password: None,
+                            token: Some(TokenAuth {
+                                id: "fake_token".to_string()
+                            }),
+                        },
+                        scope: None,
+                    },
+                }
+            )
+            .await
+            .unwrap()
+        );
+    }
+
+    #[tokio::test]
+    async fn test_authenticate_request_unsupported() {
+        let config = Config::default();
+
+        let provider = Provider::mocked_builder()
+            .config(config.clone())
+            .build()
+            .unwrap();
+
+        let state = Arc::new(
+            Service::new(
+                config,
+                DatabaseConnection::Disconnected,
+                provider,
+                MockPolicyFactory::new(),
+            )
+            .unwrap(),
+        );
+
+        let rsp = authenticate_request(
+            &state,
+            &AuthRequest {
+                auth: AuthRequestInner {
+                    identity: Identity {
+                        methods: vec!["fake".to_string()],
+                        password: None,
+                        token: None,
+                    },
+                    scope: None,
+                },
             },
-        );
-        let provider = Provider::mocked_builder()
-            .assignment(assignment_mock)
-            .identity(identity_mock)
-            .resource(resource_mock)
-            .build()
-            .unwrap();
-
-        let state = Arc::new(
-            Service::new(
-                Config::default(),
-                DatabaseConnection::Disconnected,
-                provider,
-                MockPolicyFactory::new(),
-            )
-            .unwrap(),
-        );
-        let token = ProviderToken::ProjectScope(ProjectScopePayload {
-            user_id: "bar".into(),
-            project_id: "project_id".into(),
-            roles: Some(vec![ProviderRole {
-                id: "rid".into(),
-                name: "role_name".into(),
-                ..Default::default()
-            }]),
-            ..Default::default()
-        });
-
-        let api_token = Token::from_provider_token(&state, &token).await.unwrap();
-
-        assert_eq!("bar", api_token.user.id);
-        assert_eq!(Some("user_domain_id"), api_token.user.domain.id.as_deref());
-        let project = api_token.project.expect("project_scope");
-        assert_eq!(Some("project_domain_id"), project.domain.id.as_deref());
-        assert_eq!("project_id", project.id);
-        assert!(api_token.domain.is_none());
-        assert_eq!(
-            api_token.roles,
-            Some(vec![Role {
-                id: "rid".into(),
-                name: "role_name".into(),
-                ..Default::default()
-            }])
-        );
+        )
+        .await;
+        if let KeystoneApiError::Unauthorized = rsp.unwrap_err() {
+        } else {
+            panic!("Should receive Unauthorized");
+        }
     }
 }
